+++ .gitignore
@@ -5,3 +5,4 @@ __pycache__/
 *env/
 *.py[cod]
 *$py.class
+*.pytest_cache
